splitted components of temporal encoder
