+++ src/models/temporal/__init__.py
@@ -1,0 +1,1 @@
+from .encoder import TemporalEncoder
