+++ src/models/temporal/encoder.py
@@ -3,296 +3,9 @@ import torch.nn as nn
 
 from functools import partial
 
-from models.vit.mlp import MLPBlock 
-from models.transformer.encoder import EncoderBlock, Encoder
-
-class MultiInputSequential(nn.Sequential):
-    """ A custom nn.Sequential model for multiple inputs and outputs
-    """
-    def forward(self, *inputs):
-
-        for module in self._modules.values():
-            if type(inputs) == tuple:
-                inputs = module(*inputs)
-            else:
-                inputs = module(inputs)
-        return inputs
-
-class GlobalTemporalEncoder(nn.Module):
-    """ Encode a temporal sequence of global feature with self-attention
-    """
-
-    def __init__(
-        self,
-        num_heads: int,
-        hidden_dim: int,
-        mlp_dim: int,
-        norm_layer = partial(nn.LayerNorm, eps=1e-6),
-        dropout: float=0.0,
-        need_weights = False,
-    ) -> None:
-        """Initialize Global Bi-directional Temporal Encoder
-
-        Args:
-            num_heads (int): Number of heads for the temporal attention module
-            hidden_dim (int): Hidden dimension for the temporal attention module
-            mlp_dim (int): MLP dimension for the temporal attention module
-            norm_layer (torch.nn.Module, optional): Layer norm after temporal attention. 
-                Defaults to partial(nn.LayerNorm, eps=1e-6).
-            dropout (float, optional): Dropout probability. Defaults to 0.0.
-            need_weights (bool, optional): If true, return attention weights (not configured for now). 
-                Defaults to False.
-        """
-        super().__init__()
-        
-        self.ln_1 = norm_layer(hidden_dim)
-        self.self_attention = nn.MultiheadAttention(hidden_dim, num_heads, dropout=dropout, batch_first=True)
-
-        # MLP block
-        self.ln_2 = norm_layer(hidden_dim)
-        self.mlp = MLPBlock(hidden_dim, mlp_dim)
-
-        self.need_weights = need_weights # Whether to return attention weights as well
-
-    def forward(self, inputs: torch.Tensor):
-        """Perform forward pass
-
-        Args:
-            inputs (torch.Tensor): A (B, Hw, E) tensor with Hw as the history 
-                window and E as the embedding/hidden feature dimension.
-
-        Raises:
-            result (torch.Tensor): A (B, Hw, E) tensor.
-        """
-        torch._assert(inputs.dim() == 3, f"Expected (batch_size, seq_length, hidden_dim) got {inputs.shape}")
-
-        result = None
-        attention_weights = None # Needed only if self.need_weights is True for this specific Block
-
-        x0 = self.ln_1(inputs)
-        x1, attention_weights = self.self_attention(x0, x0, x0, need_weights=self.need_weights)
-        x2 = inputs + x1
-        x3 = self.ln_2(x2)
-        x4 = self.mlp(x3)
-        result = x2 + x4
-
-        
-        return result
-
-class LocalForwardTemporalEncoder(nn.Module):
-    """ Encode a temporal sequence of local features with cross-sequence-attention
-        in forward direction
-    """
-
-    def __init__(
-        self,
-        num_heads: int,
-        hidden_dim: int,
-        mlp_dim: int,
-        norm_layer = partial(nn.LayerNorm, eps=1e-6),
-        dropout: float=0.0,
-        need_weights = False,
-        reduce: bool=False
-    ) -> None:
-        """Initialize Local Forward Temporal Encoder
-
-        Args:
-            num_heads (int): Number of heads for the temporal attention module
-            hidden_dim (int): Hidden dimension for the temporal attention module
-            mlp_dim (int): MLP dimension for the temporal attention module
-            norm_layer (torch.nn.Module, optional): Layer norm after temporal attention. 
-                Defaults to partial(nn.LayerNorm, eps=1e-6).
-            dropout (float, optional): Dropout probability. Defaults to 0.0.
-            need_weights (bool, optional): If true, return attention weights (not configured for now). 
-                Defaults to False.
-            reduce (bool, optional): If true, return the propagated attended values. Else return all the attented values. 
-                Defaults to False.  
-        """
-        super().__init__()
-        
-        self.num_heads = num_heads
-        self.reduce = reduce
-        
-        self.ln_q = norm_layer(hidden_dim)
-        self.ln_kv = norm_layer(hidden_dim)
-        
-        self.attention = nn.MultiheadAttention(hidden_dim, num_heads, dropout=dropout, batch_first=True)
-        
-        self.need_weights = need_weights
-         
-        # MLP block
-        self.ln_2 = norm_layer(hidden_dim)
-        self.mlp = MLPBlock(hidden_dim, mlp_dim) 
-         
-    def forward(self, inputs: torch.Tensor):
-        """Perform forward pass
-
-        Args:
-            inputs (torch.Tensor): A (B, Hw, Nf, E) tensor with Hw as the history 
-                window, Nf as the number of local features and E as the embedding/hidden
-                feature dimension.
-
-        Raises:
-            result (torch.Tensor): A (B, Hw', Nf, E) tensor if reduce is False else a (B, Nf, E) tensor.
-        """
-        torch._assert(inputs.dim() == 4, f"Expected Local Features of shape \
-            (batch_size, seq_length, num_feature, hidden_dim) got {inputs.shape}")
-
-        
-        ######################################################################################
-        # TODO: Need to implement a forward method for local forward temporal attention block
-        
-        b, hw, nf, e = inputs.shape 
-        # Temporal attention on local features
-        attended_values = []
-        
-        # Use oldest feature sequence from history as the first query
-        query = inputs[:, 0, :, :]        
-        
-        for i in range(1, hw-1):
-            
-            # Update query
-            key_value = inputs[:, i+1, :, :]
-
-            # Layer norm 
-            query_ln = self.ln_q(query)
-            key_value_ln = self.ln_kv(key_value)
-            
-            # Forward temporal attention on two subsequent local features at different timestep                   
-            attended_value, attention_weights = self.attention(query_ln, key_value_ln, key_value_ln)
-            
-            #############################################################################################
-            # TODO: Think about different ways we could implement/propagate attended values in such a 
-            # forward attention mechanism.
-            # - I though maybe it makes sense to add a residual after getting the attended values with 
-            #   with the original key/value and then use it as a query in the next timestep  
-            #############################################################################################
-            
-            # Update query
-            query = key_value + attended_value
-            
-            # Append attended queries
-            attended_values.append(query.unsqueeze(1))
-        
-        if self.reduce:
-            x3 = self.ln_2(query)
-            x4 = self.mlp(x3)
-            result = query + x4
-              
-        else:
-            result = torch.cat(attended_values, axis=1)
-            result = self.ln_2(result)
-            
-          
-        return result
-
-
-class LocalBackwardTemporalEncoder(nn.Module):
-    """ Encode a temporal sequence of local features with cross-sequence-attention
-        in backward direction
-    """
-
-    def __init__(
-        self,
-        num_heads: int,
-        hidden_dim: int,
-        mlp_dim: int,
-        norm_layer = partial(nn.LayerNorm, eps=1e-6),
-        dropout: float=0.0,
-        need_weights = False,
-        reduce: bool=False
-    ) -> None:
-        """Initialize Local Backward Temporal Encoder
-
-        Args:
-            num_heads (int): Number of heads for the temporal attention module
-            hidden_dim (int): Hidden dimension for the temporal attention module
-            mlp_dim (int): MLP dimension for the temporal attention module
-            norm_layer (torch.nn.Module, optional): Layer norm after temporal attention. 
-                Defaults to partial(nn.LayerNorm, eps=1e-6).
-            dropout (float, optional): Dropout probability. Defaults to 0.0.
-            need_weights (bool, optional): If true, return attention weights (not configured for now). 
-                Defaults to False.
-            reduce (bool, optional): If true, return the propagated attended values. Else return all the attented values. 
-                Defaults to False.  
-        """
-        super().__init__()
-        self.num_heads = num_heads
-        self.reduce = reduce
-        
-        self.ln_q = norm_layer(hidden_dim)
-        self.ln_kv = norm_layer(hidden_dim)
-        
-        self.attention = nn.MultiheadAttention(hidden_dim, num_heads, dropout=dropout, batch_first=True)
-        
-        self.need_weights = need_weights
-         
-        # MLP block
-        self.ln_2 = norm_layer(hidden_dim)
-        self.mlp = MLPBlock(hidden_dim, mlp_dim) 
-         
-    def forward(self, inputs: torch.Tensor):
-        """Perform forward pass
-
-        Args:
-            inputs (torch.Tensor): A (B, Hw, Nf, E) tensor with Hw as the history 
-                window, Nf as the number of local features and E as the embedding/hidden
-                feature dimension.
-
-        Returns:
-            result (torch.Tensor): A (B, Hw', Nf, E) tensor if reduce is False else a (B, Nf, E) tensor.
-        """
-        torch._assert(inputs.dim() == 4, f"Expected Local Features of shape \
-            (batch_size, seq_length, num_feature, hidden_dim) got {inputs.shape}")
-
-        
-        #######################################################################################
-        # TODO: Need to implement a forward method for local backward temporal attention block
-        
-        b, hw, nf, e = inputs.shape
-
-        # Temporal attention on local features
-        attended_values = []
-        
-        # Use oldest feature sequence from history as the first query
-        query = inputs[:, -1, :, :]        
-        
-        for i in range(hw-1, 1):
-            
-            # Update key and value
-            key_value = inputs[:, i-1, :, :]
-
-            # Layer norm 
-            query_ln = self.ln_q(query)
-            key_value_ln = self.ln_kv(key_value)
-            
-            # Backward temporal attention on two subsequent local features at different timestep                   
-            attended_value, attention_weights = self.attention(query_ln, key_value_ln, key_value_ln)
-            
-            #############################################################################################
-            # TODO: Think about different ways we could implement/propagate attended values in such a 
-            # Backward attention mechanism.
-            # - I though maybe it makes sense to add a residual after getting the attended values with 
-            #   with the original key/value and then use it as a query in the next timestep  
-            #############################################################################################
-            
-            # Update query
-            query = key_value + attended_value
-            
-            # Append attended queries
-            attended_values.append(query.unsqueeze(1))
-        
-        if self.reduce:
-            x3 = self.ln_2(query)
-            x4 = self.mlp(x3)
-            result = query + x4
-              
-        else:
-            result = torch.cat(attended_values, axis=1)
-            result = self.ln_2(result)
-            
-          
-        return result
+from models.temporal.sequential import MultiInputSequential
+from models.temporal.global_attention import GlobalTemporalAttention
+from models.temporal.local_attention import LocalBackwardTemporalAttention, LocalForwardTemporalAttention
 
       
 class TemporalEncoderBlock(nn.Module):
@@ -332,7 +45,7 @@ class TemporalEncoderBlock(nn.Module):
         self.direction = direction
         self.need_weights = need_weights 
         
-        self.local_forward_encoder = LocalForwardTemporalEncoder(
+        self.local_forward_attention = LocalForwardTemporalAttention(
             num_heads=num_heads, 
             hidden_dim=hidden_dim,
             mlp_dim=mlp_dim, 
@@ -342,7 +55,7 @@ class TemporalEncoderBlock(nn.Module):
             reduce=self.reduce
         )
         
-        self.local_backward_encoder = LocalBackwardTemporalEncoder(
+        self.local_backward_attention = LocalBackwardTemporalAttention(
             num_heads=num_heads, 
             hidden_dim=hidden_dim,
             mlp_dim=mlp_dim, 
@@ -352,7 +65,7 @@ class TemporalEncoderBlock(nn.Module):
             reduce=self.reduce
         )
         
-        self.global_encoder = GlobalTemporalEncoder(
+        self.global_attention = GlobalTemporalAttention(
             num_heads=num_heads, 
             hidden_dim=hidden_dim,
             mlp_dim=mlp_dim, 
@@ -385,11 +98,11 @@ class TemporalEncoderBlock(nn.Module):
         # TODO: Need to think about forward-backward or backward-forward methods
         
         if self.direction == "forward":
-            local_result = self.local_forward_encoder(local_feat)
+            local_result = self.local_forward_attention(local_feat)
         elif self.direction == "backward":
-            local_result = self.local_backward_encoder(local_feat)
+            local_result = self.local_backward_attention(local_feat)
 
-        global_result = self.global_encoder(global_feat)
+        global_result = self.global_attention(global_feat)
         
 
         return local_result, global_result
@@ -424,6 +137,7 @@ class TemporalEncoder(nn.Module):
             need_weights (bool, optional): If true, return attention weights (not configured for now). 
                 Defaults to False.
         """
+        torch._assert(num_layers == len(directions), "length of directions must be equal to the num_layers")
         super().__init__()
         
         self.hidden_dim = hidden_dim
