+++ src/models/temporal/sequential.py
@@ -1,0 +1,16 @@
+import torch.nn as nn
+
+
+
+
+class MultiInputSequential(nn.Sequential):
+    """ A custom nn.Sequential model for multiple inputs and outputs
+    """
+    def forward(self, *inputs):
+
+        for module in self._modules.values():
+            if type(inputs) == tuple:
+                inputs = module(*inputs)
+            else:
+                inputs = module(inputs)
+        return inputs
